requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
cachetools==5.3.2
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
coinbase-commerce==1.0.1
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
Django==3.2.1
django-allauth==0.58.1
django-ckeditor==6.6.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.1.0
django-crispy-forms==2.0
django-cron==0.6.0
django-database-url==1.0.3
django-environ==0.10.0
django-extensions==3.2.3
django-htmx==1.17.2
django-js-asset==2.1.0
django-redis==5.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
idna==3.4
kombu==5.3.5
lxml==5.1.0
more-itertools==9.1.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.2.1
Pillow==10.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.6
pyarrow==15.0.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redbird==0.7.1
redis==5.0.1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rocketry==2.5.1
selenium==4.18.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stripe==9.0.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.5.0
wsproto==1.2.0
